Enum aws_sdk_cloudcontrol::Error

#[non_exhaustive]
pub enum Error {

    AlreadyExistsException(AlreadyExistsException),
    ClientTokenConflictException(ClientTokenConflictException),
    ConcurrentModificationException(ConcurrentModificationException),
    ConcurrentOperationException(ConcurrentOperationException),
    GeneralServiceException(GeneralServiceException),
    HandlerFailureException(HandlerFailureException),
    HandlerInternalFailureException(HandlerInternalFailureException),
    InvalidCredentialsException(InvalidCredentialsException),
    InvalidRequestException(InvalidRequestException),
    NetworkFailureException(NetworkFailureException),
    NotStabilizedException(NotStabilizedException),
    NotUpdatableException(NotUpdatableException),
    PrivateTypeException(PrivateTypeException),
    RequestTokenNotFoundException(RequestTokenNotFoundException),
    ResourceConflictException(ResourceConflictException),
    ResourceNotFoundException(ResourceNotFoundException),
    ServiceInternalErrorException(ServiceInternalErrorException),
    ServiceLimitExceededException(ServiceLimitExceededException),
    ThrottlingException(ThrottlingException),
    TypeNotFoundException(TypeNotFoundException),
    UnsupportedActionException(UnsupportedActionException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlreadyExistsException(AlreadyExistsException)

The resource with the name requested already exists.

Tuple Fields of AlreadyExistsException

0: AlreadyExistsException

ClientTokenConflictException(ClientTokenConflictException)

The specified client token has already been used in another resource request.

It is best practice for client tokens to be unique for each resource operation request. However, client token expire after 36 hours.

Tuple Fields of ClientTokenConflictException

0: ClientTokenConflictException

ConcurrentModificationException(ConcurrentModificationException)

The resource is currently being modified by another operation.

Tuple Fields of ConcurrentModificationException

0: ConcurrentModificationException

ConcurrentOperationException(ConcurrentOperationException)

Another resource operation is currently being performed on this resource.

Tuple Fields of ConcurrentOperationException

0: ConcurrentOperationException

GeneralServiceException(GeneralServiceException)

The resource handler has returned that the downstream service generated an error that does not map to any other handler error code.

Tuple Fields of GeneralServiceException

0: GeneralServiceException

HandlerFailureException(HandlerFailureException)

The resource handler has failed without a returning a more specific error code. This can include timeouts.

Tuple Fields of HandlerFailureException

0: HandlerFailureException

HandlerInternalFailureException(HandlerInternalFailureException)

The resource handler has returned that an unexpected error occurred within the resource handler.

Tuple Fields of HandlerInternalFailureException

0: HandlerInternalFailureException

InvalidCredentialsException(InvalidCredentialsException)

The resource handler has returned that the credentials provided by the user are invalid.

Tuple Fields of InvalidCredentialsException

0: InvalidCredentialsException

InvalidRequestException(InvalidRequestException)

The resource handler has returned that invalid input from the user has generated a generic exception.

Tuple Fields of InvalidRequestException

0: InvalidRequestException

NetworkFailureException(NetworkFailureException)

The resource handler has returned that the request could not be completed due to networking issues, such as a failure to receive a response from the server.

Tuple Fields of NetworkFailureException

0: NetworkFailureException

NotStabilizedException(NotStabilizedException)

The resource handler has returned that the downstream resource failed to complete all of its ready-state checks.

Tuple Fields of NotStabilizedException

0: NotStabilizedException

NotUpdatableException(NotUpdatableException)

One or more properties included in this resource operation are defined as create-only, and therefore cannot be updated.

Tuple Fields of NotUpdatableException

0: NotUpdatableException

PrivateTypeException(PrivateTypeException)

Cloud Control API has not received a valid response from the resource handler, due to a configuration error. This includes issues such as the resource handler returning an invalid response, or timing out.

Tuple Fields of PrivateTypeException

0: PrivateTypeException

RequestTokenNotFoundException(RequestTokenNotFoundException)

A resource operation with the specified request token cannot be found.

Tuple Fields of RequestTokenNotFoundException

0: RequestTokenNotFoundException

ResourceConflictException(ResourceConflictException)

The resource is temporarily unavailable to be acted upon. For example, if the resource is currently undergoing an operation and cannot be acted upon until that operation is finished.

Tuple Fields of ResourceConflictException

0: ResourceConflictException

ResourceNotFoundException(ResourceNotFoundException)

A resource with the specified identifier cannot be found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ServiceInternalErrorException(ServiceInternalErrorException)

The resource handler has returned that the downstream service returned an internal error, typically with a 5XX HTTP status code.

Tuple Fields of ServiceInternalErrorException

0: ServiceInternalErrorException

ServiceLimitExceededException(ServiceLimitExceededException)

The resource handler has returned that a non-transient resource limit was reached on the service side.

Tuple Fields of ServiceLimitExceededException

0: ServiceLimitExceededException

ThrottlingException(ThrottlingException)

The request was denied due to request throttling.

Tuple Fields of ThrottlingException

0: ThrottlingException

TypeNotFoundException(TypeNotFoundException)

The specified extension does not exist in the CloudFormation registry.

Tuple Fields of TypeNotFoundException

0: TypeNotFoundException

UnsupportedActionException(UnsupportedActionException)

The specified resource does not support this resource operation.

Tuple Fields of UnsupportedActionException

0: UnsupportedActionException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>