Module aws_sdk_cloudcontrol::error

Errors that can occur when calling the service.

Modules

already_exists_exception

See AlreadyExistsException

client_token_conflict_exception

See ClientTokenConflictException

concurrent_modification_exception

See ConcurrentModificationException

concurrent_operation_exception

See ConcurrentOperationException

general_service_exception

See GeneralServiceException

handler_failure_exception

See HandlerFailureException

handler_internal_failure_exception

See HandlerInternalFailureException

invalid_credentials_exception

See InvalidCredentialsException

invalid_request_exception

See InvalidRequestException

network_failure_exception

See NetworkFailureException

not_stabilized_exception

See NotStabilizedException

not_updatable_exception

See NotUpdatableException

private_type_exception

See PrivateTypeException

request_token_not_found_exception

See RequestTokenNotFoundException

resource_conflict_exception

See ResourceConflictException

resource_not_found_exception

See ResourceNotFoundException

service_internal_error_exception

See ServiceInternalErrorException

service_limit_exceeded_exception

See ServiceLimitExceededException

throttling_exception

See ThrottlingException

type_not_found_exception

See TypeNotFoundException

unsupported_action_exception

See UnsupportedActionException

Structs

AlreadyExistsException

The resource with the name requested already exists.

CancelResourceRequestError

Error type for the CancelResourceRequest operation.

ClientTokenConflictException

The specified client token has already been used in another resource request.

ConcurrentModificationException

The resource is currently being modified by another operation.

ConcurrentOperationException

Another resource operation is currently being performed on this resource.

CreateResourceError

Error type for the CreateResource operation.

DeleteResourceError

Error type for the DeleteResource operation.

GeneralServiceException

The resource handler has returned that the downstream service generated an error that does

GetResourceError

Error type for the GetResource operation.

GetResourceRequestStatusError

Error type for the GetResourceRequestStatus operation.

HandlerFailureException

The resource handler has failed without a returning a more specific error code. This can

HandlerInternalFailureException

The resource handler has returned that an unexpected error occurred within the resource

InvalidCredentialsException

The resource handler has returned that the credentials provided by the user are

InvalidRequestException

The resource handler has returned that invalid input from the user has generated a generic

ListResourceRequestsError

Error type for the ListResourceRequests operation.

ListResourcesError

Error type for the ListResources operation.

NetworkFailureException

The resource handler has returned that the request could not be completed due to

NotStabilizedException

The resource handler has returned that the downstream resource failed to complete all of

NotUpdatableException

One or more properties included in this resource operation are defined as create-only, and

PrivateTypeException

Cloud Control API has not received a valid response from the resource handler, due to a

RequestTokenNotFoundException

A resource operation with the specified request token cannot be found.

ResourceConflictException

The resource is temporarily unavailable to be acted upon. For example, if the resource is

ResourceNotFoundException

A resource with the specified identifier cannot be found.

ServiceInternalErrorException

The resource handler has returned that the downstream service returned an internal error,

ServiceLimitExceededException

The resource handler has returned that a non-transient resource limit was reached on the

ThrottlingException

The request was denied due to request throttling.

TypeNotFoundException

The specified extension does not exist in the CloudFormation registry.

UnsupportedActionException

The specified resource does not support this resource operation.

UpdateResourceError

Error type for the UpdateResource operation.

Enums

CancelResourceRequestErrorKind

Types of errors that can occur for the CancelResourceRequest operation.

CreateResourceErrorKind

Types of errors that can occur for the CreateResource operation.

DeleteResourceErrorKind

Types of errors that can occur for the DeleteResource operation.

GetResourceErrorKind

Types of errors that can occur for the GetResource operation.

GetResourceRequestStatusErrorKind

Types of errors that can occur for the GetResourceRequestStatus operation.

ListResourceRequestsErrorKind

Types of errors that can occur for the ListResourceRequests operation.

ListResourcesErrorKind

Types of errors that can occur for the ListResources operation.

UpdateResourceErrorKind

Types of errors that can occur for the UpdateResource operation.